requirements.txt
pyyaml
qiskit>=0.44.0; extra == "quantum"
